refactor(core): split finalize, geometry builders, and bind helpers
Extract layout vs geometry/scene finalize phases; render-model lifecycle;
boxplot per-row layout; rect slot resolution; errorbar half-width; bind
field/color helpers; baseline domain training; and candidate series
lineage helpers. Facades keep import paths stable. Add characterization
for makeErrorbarHalfWidth.
