+++ packages/core/src/pipeline/bind-layer-stat-columns.ts
@@ -1,0 +1,12 @@
+/**
+ * y-channel { stat } columns each stat exposes (module-header contracts).
+ */
+export const STAT_Y_COLUMNS: Record<string, readonly string[]> = {
+  identity: [],
+  count: ["count"],
+  bin: ["count", "density", "ncount", "ndensity"],
+  density: ["density", "count", "scaled", "ndensity"],
+  smooth: [],
+  boxplot: [],
+  summary: [],
+};
