+++ packages/core/src/pipeline/compute-domains.ts
@@ -3,10 +3,11 @@
  */
 import type { PositionScale } from "../scales/train.js";
 
+import { computeTrainedBaselineDomains } from "./compute-domains-baseline.js";
 import type { FacetPanelDef } from "./facets.js";
 import { scaleDomainSnapshot } from "./layout-helpers.js";
-import { collectAxisInputs, trainAxis } from "./scale-training.js";
-import type { Advisory, LayerFrame, ScaleDomainSnapshot } from "./types.js";
+import { collectAxisInputs } from "./scale-training.js";
+import type { LayerFrame, ScaleDomainSnapshot } from "./types.js";
 import type { RunOptions } from "./types.js";
 
 export function computeEffectiveDomains(
@@ -35,47 +36,9 @@ export function computeBaselineDomains(input: {
   yInputs: ReturnType<typeof collectAxisInputs>;
   effectiveDomains: ScaleDomainSnapshot;
 }): ScaleDomainSnapshot {
-  const { options, freeX, freeY, facetPanels, panelFrames, xInputs, yInputs, effectiveDomains } =
-    input;
+  const { options, effectiveDomains } = input;
 
   if (options.baselineDomains !== undefined) return options.baselineDomains;
   if (options.baselineScales === undefined) return effectiveDomains;
-
-  const baselineX = trainAxis("x", xInputs, options.baselineScales.x).scale;
-  const baselineY = trainAxis("y", yInputs, options.baselineScales.y).scale;
-  const baselinePanels = facetPanels.map((_, panelIndex) => {
-    let x = baselineX;
-    let y = baselineY;
-    const scratch: Advisory[] = [];
-    if (freeX) {
-      const inputs = collectAxisInputs(
-        "x",
-        panelFrames[panelIndex]!,
-        options.baselineScales?.x?.type,
-        scratch,
-      );
-      x = trainAxis("x", inputs, {
-        ...options.baselineScales?.x,
-        type: baselineX.type,
-      }).scale;
-    }
-    if (freeY) {
-      const inputs = collectAxisInputs(
-        "y",
-        panelFrames[panelIndex]!,
-        options.baselineScales?.y?.type,
-        scratch,
-      );
-      y = trainAxis("y", inputs, {
-        ...options.baselineScales?.y,
-        type: baselineY.type,
-      }).scale;
-    }
-    return Object.freeze({ x: scaleDomainSnapshot(x), y: scaleDomainSnapshot(y) });
-  });
-  return Object.freeze({
-    x: scaleDomainSnapshot(baselineX),
-    y: scaleDomainSnapshot(baselineY),
-    panels: Object.freeze(baselinePanels),
-  });
+  return computeTrainedBaselineDomains(input);
 }
