+++ packages/core/src/pipeline/geometry-rects.ts
@@ -7,6 +7,7 @@ import type { LayerFrame, PipelineWarning, ResolvedColorScale } from "./types.js
 import { colorOf } from "./types.js";
 import type { Frame } from "./geometry-shared.js";
 import { DEFAULT_BAR_WIDTH, removedWarning } from "./geometry-shared.js";
+import { resolveRectSlot } from "./geometry-rects-slot.js";
 
 export function rectsBatch(
   frame: LayerFrame,
@@ -27,42 +28,19 @@ export function rectsBatch(
   const keptRows: number[] = [];
   let removed = 0;
   for (let row = 0; row < n; row++) {
+    const slot = resolveRectSlot({ frame, fx, row, binned, widthFrac });
+    if (slot === null) {
+      removed++;
+      continue;
+    }
     const t0 = fx.yScale.type === "band" ? NaN : fx.yScale.normalize(frame.ymin[row]!);
     const t1 = fx.yScale.type === "band" ? NaN : fx.yScale.normalize(frame.ymax[row]!);
-    let center: number;
-    let w: number;
-    if (binned) {
-      // Bin rects span [xmin, xmax] on a continuous x (histograms).
-      if (fx.xScale.type === "band") {
-        removed++;
-        continue;
-      }
-      const tx0 = fx.xScale.normalize(frame.xmin![row]!);
-      const tx1 = fx.xScale.normalize(frame.xmax![row]!);
-      if (Number.isNaN(tx0) || Number.isNaN(tx1) || Number.isNaN(t0) || Number.isNaN(t1)) {
-        removed++;
-        continue;
-      }
-      center = (tx0 + tx1) / 2;
-      w = Math.abs(tx1 - tx0);
-    } else {
-      const tc =
-        fx.xScale.type === "band" ? fx.xScale.normalize(frame.xValues?.[row] ?? null) : NaN;
-      if (tc === undefined || Number.isNaN(tc) || Number.isNaN(t0) || Number.isNaN(t1)) {
-        removed++;
-        continue;
-      }
-      center = tc;
-      w = widthFrac;
+    if (Number.isNaN(t0) || Number.isNaN(t1)) {
+      removed++;
+      continue;
     }
-    if (frame.dodgeSlot !== null && frame.dodgeSlotCounts !== null) {
-      const slotCount = Math.max(1, frame.dodgeSlotCounts[row]!);
-      const full = w;
-      w = full / slotCount;
-      center = center + full * ((frame.dodgeSlot[row]! + 0.5) / slotCount - 0.5);
-    }
-    const xPx = (center - w / 2) * fx.innerWidth;
-    const wPx = w * fx.innerWidth;
+    const xPx = (slot.center - slot.w / 2) * fx.innerWidth;
+    const wPx = slot.w * fx.innerWidth;
     const y0 = fx.innerHeight - t0 * fx.innerHeight;
     const y1 = fx.innerHeight - t1 * fx.innerHeight;
     rects.push(xPx, Math.min(y0, y1), wPx, Math.abs(y1 - y0));
